missed one

+++ lib/src/native/rpi_i2c_ext.cc
@@ -224,7 +224,7 @@ void serialPuts(Dart_NativeArguments arguments)
   Dart_Handle arg1 = HandleError(Dart_GetNativeArgument(arguments, 1));
   Dart_Handle arg2 = HandleError(Dart_GetNativeArgument(arguments, 2));
 
-  int fd;
+  int64_t fd;
   char *s;
   HandleError(Dart_IntegerToInt64(arg1, &fd));
   //TODO: check that this is the right dart/c conversion method
